Add tests that evaluate the Typed AST
These tests ensure that the type checking process, which creates A-normal form
DTrace programs, doesn't break the program at all.  After the probes are put
into the Typed AST, we evaluate them against initial states (of all zeros).

The new test cases define a set of probes in a .d file.  The expected results
specify a trace of probe invocations to evaluate on the initial global state.

- This change includes a modification to the `WriteGlobal` primitive to allow it
to copy values from one global to another
- It also modifies the expression translator to return `Reg` instead of local
indices, since it isn't guaranteed to refer to a local (it could be a global)