Add an LLVM code generator for DTrace probes
It is incomplete, but generates code that works for limited tests. The code generator is based on llvm-hs and llvm-hs-pure. The tests for this code generate machine code and use a runner (probe-runner.c) to load the generated code and run randomly-generated sequences of probe invocations. The result of those runs are compared against the oracle provided by the Haskell probe interpreter in Language.DTrace.

The CI has been updated to run these tests (see the locale changes required).