Inject probes into binaries using binary rewriting
The command line tool now has a subcommand for instrumenting binaries. The command determines the architecture for LLVM based on the architecture of the input binary. This commit includes the necessary calling sequences to invoke probes from the program. It also includes setup code (for x86) to set up the necessary shared memory segment to persist probe information to disk.

Currently, it only works for very simple binaries. Upcoming changes will modify the rewriting strategy to be more robust.