Rework the instrumentation strategy
This changes both the setup code and probe insertion sequence.

The setup code is now injected as new code that jumps to the original entry
point (and the entry point of the binary is modified to point to the new code).
This required some additional support in renovate.  This was necessary because
renovate is unable to rewrite the entry block of most programs because macaw
classification is a bit fragile around the unusual tail call structure of
program startup in most C libraries.  This prevents us from knowing that it is
safe to do the rewriting.  The new approach avoids any analysis of the entry
point code and forcibly inserts the required code.

The probe provider has been changed to insert probes before relevant calls,
rather than at the entry point of the callee.  This would be required to support
dynamically linked binaries anyway.  However, it is also really useful
because (similar to the setup code case), renovate is often unable to instrument
system call wrappers in libc because they have some interesting tail call
structure that renovate is too conservative to rewrite for now.  Adding the
rewriting at the call sites works around this in most cases (call sites may also
be un-instrumentable for other reasons).  The disadvantage is that calls to
target functions through function pointers will not be instrumented (but this is
unavoidable for static binaries anyway).

This commit includes a manual test (for now), as well as submodule updates.