First cut of a dockerized version of `mctrace`.
The resulting image includes a binary version of `mctrace` as well
as an installation of `musl-gcc` to allow self-contained building,
instrumentation and extraction of telemetry data.

The Dockerfile is multi-staged with the base stage installing all
prereqs and building the mctrace tool and the final stage copying
over a minimal set of built binaries to the final image to ensure a
somewhat smaller size. (Though more work can be done in this regard).