Preliminary support for injecting the runtime support library.
Currently, this requires the library to be a set of independent
functions that have no dependencies on other (external or internal)
functions. NOTE: The no-dependency requirements will likely be relaxed
in the future.

An implementation of the support library must currently contain at
least two functions: `alloc_memory` and `send`, the exact signatures
are still in flux. A simple support library, intended for testing on
X86, can be found at `mctrace/tests/library/X86/runtime.c`