Allow probes to access (specific) functions during operation (*)
This commit sets up compiled probes to access a selected set of
external functions, by passing it a pointer to a table of functions with
a well-known order (much like a vtable).

(*) This commit is not fully self-contained since it needs changes in
Renovate to work. That will be pushed as a separate commit on this
branch.