Automate the environment setup and build processes
This commit adds two new scripts that we ought to keep updated as
we move forward: dev_setup.sh (to set up a fresh install of Ubuntu
to get to a working development environment) and build.sh (to run
all steps relevant to building bits of the project).

This change also updates the README to point users at the new
scripts rather than repeating build steps there.