Add and use a platform API header
This change adds mctrace/tests/library/include/platform_api.h and
updates the two platform API implementations (runtime.c files) to use
it.

This change revealed that the two implementations differed slightly
where they differed in whether they provided a file name argument to
send(). This patch includes a fix to align them even if we ultimately
plan to remove that argument.

This change updates the 'full' test Makefile to also build the platform
API implementation for whatever platform it's building executables for.