Namespace all Mctrace platform API functions with a "platform_" prefix
This change is aimed at disambiguating between Dtrace actions (e.g.
"send()") and their underlying platform equivalents (e.g. "send()" in
C). This change adds a prefix to make the platform API implementations
easier to talk about (e.g. in documentation) and also helps to
make clear that while a Dtrace action might result in a call to a
similarly-named platform function, they are not the same thing.