language-dtrace: typeOfExpr returns Nothing for the guessed type of Calls
The language-dtrace type checker allocates implicit global variables
and sets their types based on the right-hand side expression type in an
initial assignment. So, if we have

    x = foo()

and 'x' has not already been explicitly declared a global, this
assignment (as the first mention of 'x') will be used to indicate that
'x' is a global and the return type of foo() will be used as x's type.
However, the type of the RHS is obtained from foo() as an *untyped* AST
node, i.e., it has not undergone rewriting as a consequence of type
checking. Since calls like "foo()" are represented opaquely in the
untyped AST -- that is, the AST indicates only the name and arguments of
the call, not its return type -- we can't make a guess about the return
type without hard-coding a mapping in the type checker about the types
of the calls we support. We could do that, but that would be an ugly
hack, and besides, it isn't necessary because we can make the type guess
(performed here by typeOfExpr) return Nothing for calls. The consequence
of that is that lines of code like the one above are only valid if 'x'
has been explicitly declared as a global, and I don't think that's such
a bad thing anyway. If that became unpalatable for some reason, then
the solution is to obtain the type information some other way, such as
by having a hard-coded mapping or by consulting a future procedure call
lookup table generically.