symAddressForSymbolPattern: improve performance by only doing map filtering when necessary
Prior to this change, symAddressForSymbolPattern always did a map filter
operation to find matching symbols regardless of the probe component
that was being matched. That meant that fetching the list of relevant
symbols was always O(n) for the size of the map, even though in the vast
majority of cases, a single map key lookup was needed. This change
updates symAddressForSymbolPattern to do a map filter only for
pattern-based matching, a no-op for the free pattern, and a single map
key lookup for the case where we need to match an exact identifier. This
led to approximately 90% reduced runtime over the previous
implementation for a 4 MB binary that I used for testing.