Update code generation and rewriting to support `arg0`.
Note that the value of `arg0` differs in the entry and return probes. In
the former it is the value of the first argument to the function being
probed, while in the latter it is the return value from the function.

Probes now take an additional argument representing the value of `arg0`.
This is expected to be the 3rd argument, pushing `ucaller` value to be
the 4th argument to the probe. The code generator has been updated to
handle this.

The probe providers for X86-64 and PPC have been updated to generate
instructions to pick out the correct value of `arg0` (based on calling
convention for the architecture and type of the probe) prior to
invoking the probe.