Split up the documentation along functionality and packaging lines
This change does a few things aimed at making it easier to maintain our
documentation:

* Moves much of the introductory/explanatory content from the release
  README (release/README.md) into a new document specific to the mctrace
  tool itself, MCTRACE.md.

* Updates the main README.md to focus more on developer-oriented
  instructions for doing a local build and producing release images.

* Updates the release README to refer to MCTRACE.md rather than
  repeating much of what it now contains.

Split this way, each document now has a (to me) much clearer purpose and
the risk of duplication is much lower as a result. It also means that
each document has a pretty clear, single audience rather than needing to
be written to be useful for both developers and release consumers or end
users.