Initial version of mctrace as a dockerized command.
This commit includes a new Dockerfile-for-tool that builds on
the original release Dockerfile to create a "dockerized" mctrace
command. The image will only contain the `mctrace` tool and
associated dependencies. The command is set as the `ENTRYPOINT`
to enable it to be run directly.

It also includes a script (also called `mctrace`) that runs the
container with volumes mounted such that paths relative to the repo
on the host are aligned with the default working directory in the
container. This allows the dockerized command to be invoked with
files/paths relative to the repo and have it work "correctly".