Implement a really basic elf-patching mechanism.
The concern at hand is that the binary analysis framework at large makes
assumptions about the structure of ELF files it deals with. For example,
some files may have their `.plt` section tagged as `NOBITS` (which seems
perfectly valid for PPC), but that will interact badly with the
assumptions in the rewriting framework. Where fixing the latter is
difficult or we are constrained on time, we could use the patching
mechanism to patch things over before the rest of the binary analysis
framework sees the file.

The current implementation is super rudimentary and only covers a single
case. Further, it does not handle *unpatching*, (i.e. revert the
changes we have made to the ELF), which is likely a good idea in
practice. For example, creating a bracketed API like `withPatchedELF` or
something of that sort that automatically handles this could be
beneficial.

Note that there *are* certain situations we cannot handle. If `elf-edit`
itself makes assumptions (for example, it assumes that the flags for
`.got` section should exactly be `WRITE | ALLOC`), we cannot handle that
with this framework (since we use that package to do the actual edit).