Add a new .got section flag patch, and make all patches optional and user-configurable
This change introduces another patch to set an input binary's .got
section flags to W/A, since we found a case where it can have other
flags that violate assumptions and lead to a crash.

As part of this work, since this patch and the .plt patch that was
already in place are not always desirable, this change introduces
infrastructure to make the elf-patching code configurable in that it can
be told which patches to apply. This change introduces new command-line
flags to the 'instrument' command that allow the user to enable those
patches. They are both disabled by default.

Unfortunately, it is difficult tell when a particular patch is needed
just by looking at the tool's crash output. This change doesn't do
anything to attempt to give the user a suggestion about which patch to
enable based on the failure of the tool; that would be nice, but doing
that probably requires deeper changes to some of the underlying
libraries.